requirements.txt
numpy
google-generativeai
groq
chromadb
fastapi
uvicorn[standard]
pypdf
python-docx
requests
beautifulsoup4
